+++ src/main/java/com/wenxin2/marioverse/entities/MegaGoombaEntity.java
@@ -76,6 +76,11 @@ public class MegaGoombaEntity extends GoombaEntity implements GeoEntity {
     }
 
     @Override
+    public float getVoicePitch() {
+        return (this.random.nextFloat() - this.random.nextFloat()) * 0.2F + 0.5F;
+    }
+
+    @Override
     protected void registerGoals() {
         this.goalSelector.addGoal(0, new MeleeAttackGoal(this, 0.6D, false));
         this.goalSelector.addGoal(1, new RandomStrollGoal(this, 0.4D));